pyproject.toml
[project]
name = "knowledge-collection-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pmaw>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "readability-lxml>=0.8.4.1",
    "sentence-transformers>=5.1.2",
    "transformers>=4.57.1",
    "umap-learn>=0.5.9.post2",
    "asyncpg>=0.31.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "hdbscan>=0.8.38",
    "scikit-learn>=1.6.1",
    "numpy>=2.2.1",
    "scipy>=1.15.1",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
